+++ changelog.md
@@ -66,6 +66,7 @@ Codion Change Log
 - ListSelectionDialogBuilder.comparator() added.
 - Dialogs.inputDialog() overloaded with ComponentBuilder.
 - FilterList now scrolls to the selected item.
+- FilterTableColumn.Builder.fixedWidth() added.
 ### is.codion.swing.framework.ui
 - EntityApplicationPanel.Builder.applicationModelFactory() and applicationPanelFactory() renamed applicationModel() and applicationPanel().
 - EntityApplicationPanel.createLogControls() now protected final.